Move entities, rename scheduler and add refactor comments
