+++ src/domain/scheduler/entities/Logger.ts
