+++ src/domain/scheduler/entities/ScheduledJob.ts
