+++ src/domain/scheduler/entities/SchedulerContract.ts
