+++ src/scheduler/Scheduler.ts
@@ -1,6 +1,6 @@
 import schedule from "node-schedule";
-import { SchedulerContract } from "./entities/SchedulerContract";
-import { DEFAULT_SCHEDULED_JOB_ID, ScheduledJob } from "./entities/ScheduledJob";
+import { SchedulerContract } from "../domain/scheduler/entities/SchedulerContract";
+import { DEFAULT_SCHEDULED_JOB_ID, ScheduledJob } from "../domain/scheduler/entities/ScheduledJob";
 import { EVERY_MINUTE_FREQUENCY } from "../domain/scheduler/entities/SyncRuleJobConfig";
 
 export default class Scheduler implements SchedulerContract {
