+++ src/scheduler/SchedulerCLI.ts
@@ -1,13 +1,16 @@
 import moment from "moment";
 import cronstrue from "cronstrue";
 import { CompositionRoot } from "../presentation/CompositionRoot";
-import { SchedulerContract } from "./entities/SchedulerContract";
-import { DEFAULT_SCHEDULED_JOB_ID, ScheduledJob } from "./entities/ScheduledJob";
+import { SchedulerContract } from "../domain/scheduler/entities/SchedulerContract";
+import { DEFAULT_SCHEDULED_JOB_ID, ScheduledJob } from "../domain/scheduler/entities/ScheduledJob";
 import { SyncRuleJobConfig } from "../domain/scheduler/entities/SyncRuleJobConfig";
 import { SchedulerExecutionInfo } from "../domain/scheduler/entities/SchedulerExecutionInfo";
-import { Logger } from "./entities/Logger";
+import { Logger } from "../domain/scheduler/entities/Logger";
 
-export class SchedulerPresenter {
+/**
+ * @description This file is refactored
+ */
+export class SchedulerCLI {
     constructor(
         private options: {
             scheduler: SchedulerContract;
@@ -18,18 +21,20 @@ export class SchedulerPresenter {
 
     public initialize(apiPath: string): void {
         this.fetchTask(apiPath);
+        const { scheduler, logger } = this.options;
 
-        this.options.scheduler.scheduleJob({ jobCallback: (): Promise<void> => this.fetchTask(apiPath) });
+        scheduler.scheduleJob({ jobCallback: (): Promise<void> => this.fetchTask(apiPath) });
 
-        this.options.logger.info("main", `Loading synchronization rules from remote server`);
+        logger.info("main", `Loading synchronization rules from remote server`);
     }
 
     private async fetchTask(apiPath: string): Promise<void> {
+        const { scheduler, logger } = this.options;
         try {
             const syncRuleJobConfigs = await this.getSyncRuleJobConfigs();
             const jobIdsToBeScheduled = syncRuleJobConfigs.map(({ id }) => id);
 
-            const scheduledJobs = this.options.scheduler.getScheduledJobs();
+            const scheduledJobs = scheduler.getScheduledJobs();
             const currentJobIdsScheduled = scheduledJobs.map(({ id }) => id);
 
             const jobIdsToCancel = currentJobIdsScheduled.filter(
@@ -49,30 +54,33 @@ export class SchedulerPresenter {
             this.updateNextExecutionOfScheduler(scheduledJobs);
         } catch (error) {
             const errorMessage = typeof error === "string" ? error : JSON.stringify(error, null, 2);
-            this.options.logger.error("scheduler", `${errorMessage}`);
+            logger.error("scheduler", `${errorMessage}`);
         }
     }
 
     private async synchronizationTask(ruleId: string, apiPath: string): Promise<void> {
-        const rule = await this.options.compositionRoot.rules.get(ruleId);
+        const { logger, compositionRoot } = this.options;
+
+        const rule = await compositionRoot.rules.get(ruleId);
+
         if (!rule) return;
 
         const { name, frequency, builder, id: syncRule, type = "metadata" } = rule;
 
         try {
             const readableFrequency = cronstrue.toString(frequency || "");
-            this.options.logger.debug(name, `Start ${type} rule with frequency: ${readableFrequency}`);
-            const result = await this.options.compositionRoot.sync.prepare(type, builder);
-            const sync = this.options.compositionRoot.sync[type]({ ...builder, syncRule });
+            logger.debug(name, `Start ${type} rule with frequency: ${readableFrequency}`);
+            const result = await compositionRoot.sync.prepare(type, builder);
+            const sync = compositionRoot.sync[type]({ ...builder, syncRule });
 
             const synchronize = async () => {
                 for await (const { message, syncReport, done } of sync.execute()) {
-                    if (message) this.options.logger.debug(name, message);
-                    if (syncReport) await this.options.compositionRoot.reports.save(syncReport);
+                    if (message) logger.debug(name, message);
+                    if (syncReport) await compositionRoot.reports.save(syncReport);
                     if (done && syncReport && syncReport.id) {
                         const reportUrl = this.buildUrl(apiPath, type, syncReport.id);
-                        this.options.logger.debug(name, `Finished. Report available at ${reportUrl}`);
-                    } else if (done) this.options.logger.warn(name, `Finished with errors`);
+                        logger.debug(name, `Finished. Report available at ${reportUrl}`);
+                    } else if (done) logger.warn(name, `Finished with errors`);
                 }
             };
 
@@ -84,29 +92,31 @@ export class SchedulerPresenter {
                     switch (code) {
                         case "PULL_REQUEST":
                         case "PULL_REQUEST_RESPONSIBLE":
-                            this.options.logger.error(name, "Metadata has a custodian, unable to proceed with sync");
+                            logger.error(name, "Metadata has a custodian, unable to proceed with sync");
                             break;
                         case "INSTANCE_NOT_FOUND":
-                            this.options.logger.error(name, "Couldn't connect with instance");
+                            logger.error(name, "Couldn't connect with instance");
                             break;
                         case "NOT_AUTHORIZED":
-                            this.options.logger.error(name, "User is not authorized to one or more instances");
+                            logger.error(name, "User is not authorized to one or more instances");
                             break;
                         default:
-                            this.options.logger.error(name, "Unknown synchronization error");
+                            logger.error(name, "Unknown synchronization error");
                     }
                 },
             });
         } catch (error) {
             const errorMessage = typeof error === "string" ? error : JSON.stringify(error, null, 2);
-            this.options.logger.error(name, `Failed executing rule: ${errorMessage}`);
+            logger.error(name, `Failed executing rule: ${errorMessage}`);
         }
     }
 
     private async getSyncRuleJobConfigs(): Promise<SyncRuleJobConfig[]> {
-        const syncRuleJobConfigsToBeScheduled = await this.options.compositionRoot.scheduler.getSyncRuleJobConfigs();
+        const { logger, compositionRoot } = this.options;
 
-        this.options.logger.trace(
+        const syncRuleJobConfigsToBeScheduled = await compositionRoot.scheduler.getSyncRuleJobConfigs().toPromise();
+
+        logger.trace(
             "scheduler",
             `There are ${syncRuleJobConfigsToBeScheduled.length} valid sync rules marked to be scheduled`
         );
@@ -115,21 +125,25 @@ export class SchedulerPresenter {
     }
 
     private cancelScheduledJobs(jobIdsToCancel: string[], scheduledJobs: ScheduledJob[]): void {
+        const { scheduler, logger } = this.options;
+
         jobIdsToCancel.forEach((id: string) => {
-            this.options.logger.info("scheduler", `Cancelling disabled rule with id ${id}`);
+            logger.info("scheduler", `Cancelling disabled rule with id ${id}`);
             const scheduledJobToCancel = scheduledJobs.find(scheduledJob => scheduledJob.id === id);
             if (scheduledJobToCancel) {
-                this.options.scheduler.cancelJob(scheduledJobToCancel.id);
+                scheduler.cancelJob(scheduledJobToCancel.id);
             }
         });
     }
 
     private createNewScheduledJobs(syncRuleJobConfig: SyncRuleJobConfig[], apiPath: string): void {
         syncRuleJobConfig.forEach((syncRuleJobConfig: SyncRuleJobConfig): void => {
+            const { scheduler, logger } = this.options;
+
             const { id, name, frequency } = syncRuleJobConfig;
 
             if (id && frequency) {
-                const job = this.options.scheduler.scheduleJob({
+                const job = scheduler.scheduleJob({
                     jobId: id,
                     frequency: frequency,
                     jobCallback: (): Promise<void> => this.synchronizationTask(id, apiPath),
@@ -137,15 +151,17 @@ export class SchedulerPresenter {
 
                 // Format date to keep timezone offset
                 const nextDate = moment(job.nextExecution.toISOString()).toISOString(true);
-                this.options.logger.info("scheduler", `Scheduling new sync rule ${name} (${id}) at ${nextDate}`);
+                logger.info("scheduler", `Scheduling new sync rule ${name} (${id}) at ${nextDate}`);
             }
         });
     }
 
     private async updateNextExecutionOfScheduler(scheduledJobs: ScheduledJob[]): Promise<void> {
+        const { compositionRoot } = this.options;
+
         const defaultScheduledJob = scheduledJobs.find(scheduledJob => scheduledJob.id === DEFAULT_SCHEDULED_JOB_ID);
         const nextExecution = defaultScheduledJob?.nextExecution;
-        const schedulerExecutionInfo = await this.options.compositionRoot.scheduler.getLastExecutionInfo();
+        const schedulerExecutionInfo = await compositionRoot.scheduler.getLastExecutionInfo().toPromise();
 
         const newSchedulerExecutionInfo: SchedulerExecutionInfo = {
             ...schedulerExecutionInfo,
@@ -153,7 +169,7 @@ export class SchedulerPresenter {
             nextExecution: nextExecution,
         };
 
-        await this.options.compositionRoot.scheduler.updateExecutionInfo(newSchedulerExecutionInfo);
+        await compositionRoot.scheduler.updateExecutionInfo(newSchedulerExecutionInfo).toPromise();
     }
 
     private buildUrl(apiPath: string, type: string, id: string): string {
