+++ src/scheduler/__tests__/mocks/MockLogger.ts
@@ -1,4 +1,4 @@
-import { Logger } from "../../entities/Logger";
+import { Logger } from "../../../domain/scheduler/entities/Logger";
 
 export class MockLogger implements Logger {
     trace(name: string, message: string): void {
